refactor(FloatingBox): Move content to file
